Reworked the server to not block the event thread
On the server, the while loop used to simulate the game was completely
blocking the node event loop. This meant that if a large number of
simulations were running, no socket io pings ould be processed so the
connection would timeout, and the results not displayed. To avoid this,
the non-view simulator has had a timeout added to it which lets the
event loop break from processing th esimulation to service other
requests. To ensure it is still fast at running the simulations, the
frequency of these breaks is not very high.

It also means the code that runs the simulations had to be reworked,
since the for loops would just start a number of threads that ran in
paralles. I think it's preferrable to have them one at a time, wich has
been achieved witha recursive function.

The simulation running code is now very similar for both visual and
non-visual approaches, so have been combined into a single function.