Added loading bar to simulation view
This updates the progress of the simulation so the user nows how close
it is to completing. This is useful as running a lot of simulations can
take along time. It works by recording how many simulations have
completed vs how many should complete in total.

A new oading bar react component has been added.