Further refinement and balancing of game
Have tweaked a lot of parameters, including tower properties, enemy
properties, round config and more. Also added some additional tower
purchase methods.

I have noted a few things whilst doing this process:
1. Having towers with the same DPS per cost is not a good idea. The
reason for this is that it means there is no reason to ever by a more
expensive tower. I played a very successful run with just the shrapnel
burst, and not purchasing any other tower. Yes, the map would get
crowded, but thats only if more than one player most likely. As such,
later towers will be more powerful for the price, and later enemies will
be more dangerous.

2. The best tower purchase methods seem to be the ones that try to buy
the most recently unlocked tower. This is desireable, as it means that
the more expensive towers are in fact more powerful. The "most
expensive" puchase method is different, as it buys the most expensive
tower each round, so never saves up.

3. The way the graphs are currently rendered shows that on average, some
tracks are easy, i.e. very few lives lost, and others are harder. I
think this is what we want, as I expect it will be very difficult to
have the lives gradually going down. As such, as long as a certain
percentage of the runs do ok, some completely fail, and others lose
quite a lot of lives, then it should be balanced.

I think assuming the player will be better than the AI is reasonable,
though I am currently unsure. I think I will have to do some personal
user testing to discover this. Quite possibly the player will not be as
good, especially since the AI can determine the optimal place for a
tower. Although "optimal" means most nimber of track parts in it's
range, which may not be true. AI also does not take into account things
like aim, selling, and upgrades. I likely will not make the computer do
this for simulations, and will be a lot of work to make a decent
heuristic for this. Upgrades may be ok, but certainly not selling and
aiming, that is too complex. Behaviour with multiple players may also
make difficulty completely different.