Added new enemies
Bonehead thug, which is stronger than bandit but weaker than shield.
Also added bonehead tank, which currently does not have it's own sprite,
so have temporarily removed uber slime so no sprite clash.

THus is useful because shield was proving a little to difficult, so thug
was added for better progression. Tank is strongest enemy in the game.