Further improved simulator
Added a couple of new methods of tower placement, which improve upon the
methods that allow a maximum number of purchases for each tower. Four
and five seem to be the best so far.

Also added some placement techniques other than most paths in range.
THis is used for specific towers so that they are placed a little more
optimally.

Finally, added the ability to purchase towers mid-round, as this is more
realistic. It seems to greatly improve performance resulting is fewer
lives lost.