Started work on end of game modals
This is a modal that pops up once the game has ended, either victoy,
loss, or completion. Decided to create it in react, which required some
shifting around of the exsitng react came class. In order for the modal
to scale with the size of the game canvas, have to have the size control
outside of the game canvas class. As such, the canvas element is now
passed through to the sprite handler once initially rendered.

Also requires an explicit set state call from the server once the state
of the game has changed from active.