Added button to return to main menu after game finished
This returns the player to the main menu. Because it is shown after the
game is over, it also removes the player form the game and removes the
stored socket. A loop was added to periodically check games to see if no
players are remaining. If this is the case then the game itself is
removed.

What this means for players acidentally disconnecting is that if they
all do, the game is removed. If at least one player remains, the game
will persist. Perhaps need to change this to only removing of the game
has actually finished.