Changed points/money allocation for multiple players
When there are multiple players, the money for each kill is given to all
players. This keeps things fair and ensures the winning player does not
continue to win as they can buy more towers and thus earn more money.
Points are also allocated to the player that does the most damage to an
enemy, as a sort of assist. Half points for this.