Added upgrade buying to simulation
This can be used to ensure upgrades are balanced and strong enough. This
initial work adds it on to the simulation side. In the same way as the
tower purchasing methods, there will be different upgrade buying
methods. Further changes are needed on the front end to allow selection
of uprade purchase types.