Added health bars to enemies
Underneath each enemy sprite, a small health bar has been added. I think
this is useful for a couple of reasons. Firstly, it can help with
tactical decisions, such as overall how well the towers are doing, and
if one needs to be added to help finish a round. It also gives good
feedback to a user when they upgrade or purchase a tower, as they can
see the health going down faster which is satisfying. The colours are
quite bright, but I think that is a good thing. I have also tweaked the
hit animation to be slightly transparent, increasingly so as it moves
through the frames. THis means the health bars will not be compketely
covered.
I may look into different colours/designs in the furture.