Added memoization to certain strings on menu buttons
To ensure menu button text fits, it is shrunk until criteria is met.
This is an expensive process, so to speed it up results are saved in a
map. Since common strings will be used repeatedly as towers are clicked
on, it is better to store this data than to recompute each time.