Allow game to be started if a player has left
Originally if a player left, the round could not be started since it
required all players to start. Now, only the connected players need to
start a round for it to begin.

Also, the ready status and button disabling is driven by game updates,
not socket events. THis means if a player re-joins they can see the
state of the readiness.